pyproject.toml
[project]
name = "x-search"
version = "0.1.0"
description = "AI-powered semantic search for your Twitter likes"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

dependencies = [
    # Core
    "python-dotenv>=1.0.0",

    # Database
    "psycopg2-binary>=2.9.9",
    "SQLAlchemy>=2.0.0",

    # Vector & Search
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",

    # NLP & Embeddings
    "sentence-transformers>=2.2.2",
    "torch>=2.1.2",
    "transformers>=4.36.2",

    # Web Scraping (optional - only if ENABLE_LINK_SCRAPING=true)
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "html5lib>=1.1",
    "newspaper3k>=0.2.8",
    "playwright>=1.40.0",

    # AI/LLM (at least one required for AI answers)
    "anthropic>=0.18.0",
    "openai>=1.0.0",

    # UI
    "streamlit>=1.32.0",

    # Utilities
    "loguru>=0.7.2",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 120
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src"
